requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email_validator
werkzeug
python-dotenv
pandas
openpyxl
cloudinary
